remove grpah page and move graph to program page, format wireframe for profile page
