+++ client/src/components/App.scss
@@ -22,12 +22,13 @@
   }
 
   .contents {
-    width: 375px;
+    min-width: 375px;
+    width: 100%;
     display: flex; 
      flex-direction: column;
     justify-content: space-between;
-    align-items: center;   
-    padding: 100px 40px;
+ 
+    padding: 100px 30px 100px;
     height: 100%;
   }
   
@@ -60,7 +61,7 @@
   
   .badge img {
     font-size: 20px;
-    height: 200px;
+    height: 100px;
     width: fit-content;
   }
   
@@ -80,5 +81,26 @@
   gap: 10px;
 }
 
+.current-program {
+  display: flex;
+  flex-direction: row;
+}
 
+.root-btns {
+  flex-direction: row;
+  gap: 30px;
+}
+
+.profile-header {
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  width: 100%;
+}
+
+.profile-content {
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+}
