+++ client/src/components/Button.js
@@ -1,10 +1,9 @@
-import "bulma/css/bulma.css";
 import { useNavigate } from "react-router-dom";
 
 const Button = (props) => {
   const navigate = useNavigate();
   return (
-    <button className="button is-success is-light is-large" onClick={() => {
+    <button className="button is-light " onClick={() => {
       navigate(props.path)
     }}>
       {props.word}
