+++ client/src/components/Header.js
@@ -1,11 +1,15 @@
-import React from 'react'
-import "./Header.scss"
-import logo from '../images/️toStrong-logos_black.png';
+import React from "react";
+import { Link } from 'react-router-dom';
+import "./Header.scss";
+import logo from "../images/️toStrong-logos_black.png";
 
 export default function Header() {
   return (
-    <div className='header'>
-      <img src={logo} />
+    <div className="header">
+      <Link to="/">
+        {" "}
+        <img src={logo} alt="app-logo" />{" "}
+      </Link>
     </div>
-  )
+  );
 }
