+++ client/src/components/Header.scss
@@ -2,7 +2,7 @@
   display: flex;
   flex-direction: row;
   position: fixed;
-  width: 32.4%;
+  width: 61.5%;
   align-items: center;
   justify-content: space-between;
   background-color: rgb(167, 167, 167);
