+++ client/src/pages/View.js
@@ -3,5 +3,6 @@ import React from 'react'
 export default function View() {
   return (
     <div>View</div>
+    // to view detail from exercise library
   )
 }
